+++ resources/views/manager/windows/fields/edit.blade.php
@@ -13,8 +13,8 @@
         <h3>{{ $componentTitle }}</h3>
 
         <div data-vue-fields class="space-y-8">
-            @foreach($fields->all() as $i => $fieldGroup)
-                @include('chief::manager.fields.form.fieldgroup', ['index' => $i])
+            @foreach($fields->all() as $i => $fieldSet)
+                @include('chief::manager.fields.form.fieldSet', ['index' => $i])
             @endforeach
         </div>