+++ CHANGELOG.md
@@ -4,7 +4,7 @@ principles.
 
 ## next release
 - Added: Visitable::response() which renders the page view by default but also allows for custom handling a frontend request of the model.
-- Added: FieldWindow and FieldGroup models that allow for more control on your field structure.
+- Added: FieldWindow and FieldSet models that allow for more control on your field structure.
 - Added: File thumbnail is now shown in the field window for pdf files.
 - Fixed: Link window could not be used without a State window.
 - Removed: Field::component() to select the window where this field should be placed in. You should now use the FieldWindow syntax instead.